requirements.txt
websockets
pyserial-asyncio
pyserial